refactor: pull preview page component to its own component
